project/project_env/WeatherApp/requirements.txt
asgiref==3.2.5
astroid==2.3.3
colorama==0.4.3
Django==3.0.4
get==2019.4.13
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
post==2019.4.13
public==2019.4.13
pylint==2.4.4
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
